#3359 - Reuse Docker-on-request for AWSLambda and Batch
